+++ vanadium-benches/src/bin/small-glommio.rs
@@ -3,12 +3,12 @@ extern crate openblas_src;
 use tcmalloc::TCMalloc;
 
 use vanadium_benches::{bench_covariance, small_header};
-use vanadium_io::{BackendSelector, get_image_f32};
+use vanadium_io::bip::GlommioBip;
 
 #[global_allocator]
 static GLOBAL: TCMalloc = TCMalloc;
 
 fn main() {
-    let mut image = get_image_f32(BackendSelector::Glommio, small_header()).unwrap();
-    bench_covariance(image.as_mut())
+    let mut image: GlommioBip<f32> = GlommioBip::new(small_header());
+    bench_covariance(&mut image)
 }
