+++ vanadium-io/Cargo.toml
@@ -5,7 +5,8 @@ edition = "2018"
 
 [features]
 progress = ["indicatif"]
-glommio-support = ["glommio", "futures"]
+glommio-backend = ["glommio", "futures"]
+syscall-backend = []
 
 [dependencies]
 ndarray = "^0.15.3"
